Move loader to features
