+++ tests/test_modules/test_loader.py
@@ -5,7 +5,7 @@ from tempfile import mkdtemp
 
 import pandas as pd
 
-from corebehrt.modules.loader import FormattedDataLoader
+from corebehrt.modules.features.loader import FormattedDataLoader
 
 
 class TestFormattedDataLoader(unittest.TestCase):
